Move AttributesMap to :sdk:common
